(feature) basic imports support
Generated Java code uses imports rather than fully qualified type names.

As a consequence, TypeMapper now only retrun the simple
(unqualified) type name corresponding to a Smef type sepcification.

A possible caveat with this approach is that it assumes that there
are no simple name collisions. As the assumption is true in most
cases, this will be addressed later.